requirements.txt
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
numpy==1.24.3
Pillow==10.3.0
tqdm==4.66.4
